backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
numpy==1.24.3
tensorflow==2.14.0
mediapipe==0.10.7
opencv-python==4.8.1.78
pymongo==4.5.0
python-dotenv==1.0.0
google-cloud-translate==3.12.0
google-cloud-texttospeech==2.14.1
scikit-learn==1.3.2
pandas==2.1.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
requests==2.31.0
aiofiles==23.2.1
pytest==7.4.3
